Show legend when a threshold or set value is set.
Only show the legend when the threshold or a set value is set.
When the legend is shown, the dataset itself is not listed in the
legend.

The legend is displayed under the plot, because left or right would make the plot small.

Also increased the thickness of the lines (I found it was not very
visible on my screen).